+++ core/src/test/kotlin/dev/zieger/utils/observables/ObservableTest.kt
@@ -7,10 +7,10 @@ import dev.zieger.utils.core_testing.assertion.assert
 import dev.zieger.utils.core_testing.assertion.rem
 import dev.zieger.utils.coroutines.scope.DefaultCoroutineScope
 import dev.zieger.utils.misc.asUnit
-import dev.zieger.utils.observable.IObservable2
+import dev.zieger.utils.observable.IObservableWithParent
 import dev.zieger.utils.observable.Observable
 import dev.zieger.utils.observable.ObservableParams
-import dev.zieger.utils.observable.ObservableParams2
+import dev.zieger.utils.observable.ObservableParamsWithParent
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.sync.Mutex
@@ -27,7 +27,7 @@ class ObservableTest {
         param("notifyForExisting", true, false),
         param("notifyOnChangedOnly", true, false),
         param("scope", DefaultCoroutineScope(), null),
-        param("mutex", Mutex(), null),
+        param("mutex", Mutex()),
         block = factory
     )
 
@@ -36,7 +36,7 @@ class ObservableTest {
         TestClass("foo", it) { initial, notifyOnChangedOnly, notifyForExisting,
                                storePreviousValues, scope, mutex ->
             Observable(
-                ObservableParams2(
+                ObservableParamsWithParent(
                     initial, notifyOnChangedValueOnly = notifyOnChangedOnly,
                     notifyForInitial = notifyForExisting,
                     storeRecentValues = storePreviousValues, scope = scope, mutex = mutex
@@ -62,7 +62,7 @@ class ObservableTest {
         }
     }
 
-    private fun <O : IObservable2<Any?, String>> runTestForInputFactory(
+    private fun <O : IObservableWithParent<Any?, String>> runTestForInputFactory(
         inputFactory: (Map<String, ParamInstance<*>>) -> TestClass<String, O>
     ) = runTest {
         params(inputFactory) {
@@ -76,10 +76,10 @@ class ObservableTest {
             testClass.observable.observe { /*Log.d("observe -> value=$it");*/ latestObservedChanged = it }
             latestObservedChanged assert (if (notifyForExisting) "foo" else null) % "after observe without changes"
 
-            var latestObservedChangedS: String? = null
-            testClass.observable.observeS { /*Log.d("observeS -> value=$it");*/ latestObservedChangedS = it }
-            delay(10L)
-            latestObservedChangedS assert (if (notifyForExisting && scope != null) "foo" else null) % "after observe without changes suspended"
+//            var latestObservedChangedS: String? = null
+//            testClass.observable.observeS { /*Log.d("observeS -> value=$it");*/ latestObservedChangedS = it }
+//            delay(10L)
+//            latestObservedChangedS assert (if (notifyForExisting && scope != null) "foo" else null) % "after observe without changes suspended"
 
             testClass.value = "boo"
             delay(10L)
@@ -87,23 +87,23 @@ class ObservableTest {
             testClass.observable.value assert "boo" % "after change observable.value"
             testClass.value assert "boo" % "after change value"
             latestObservedChanged assert "boo" % "after change observe"
-            latestObservedChangedS assert (if (scope != null) "boo" else null) % "after change observe suspended"
+//            latestObservedChangedS assert (if (scope != null) "boo" else null) % "after change observe suspended"
         }
     }.asUnit()
 
     companion object {
 
-        data class TestClass<T, O : IObservable2<Any?, T>>(
+        data class TestClass<T, O : IObservableWithParent<Any?, T>>(
             val initial: String,
             val map: Map<String, ParamInstance<*>>,
-            val factory: (String, Boolean, Boolean, Boolean, CoroutineScope?, Mutex?) -> O
+            val factory: (String, Boolean, Boolean, Boolean, CoroutineScope?, Mutex) -> O
         ) {
 
             val storePreviousValues: Boolean by bind(map)
             val notifyForExisting: Boolean by bind(map)
             val notifyOnChangedOnly: Boolean by bind(map)
             val scope: CoroutineScope? by bind(map)
-            val mutex: Mutex? by bind(map)
+            val mutex: Mutex by bind(map)
 
             val observable = factory(
                 initial, notifyOnChangedOnly, notifyForExisting,