+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+##### 2.2.43
+
+* multiple non released version because of JitPack foo
+* fix for `suspendUntil` in `OnChanged` and `Observable`
+* changed the name of `OnChanged2` to `OnChangedWithParent` (same for `ObservableWithParent`) 
+
 ##### 2.2.38
 
 *  FiFo crash fix (add new value before removing old ones)
@@ -76,7 +82,7 @@
 ##### 2.2.22
 
 * Continuation:
-    * allow multiple `suspendUntilTrigger` calls for the same trigger event
+    * allow multiple `suspend` calls for the same trigger event
     * made coroutine property nullable
     * never launch a coroutine for unsuspended trigger call when no coroutine was defined